A bit of cleanup (#761)
